requirements.txt
Django==4.2.7
Pillow==10.1.0
torch==2.2.0
torchvision==0.17.0
transformers==4.36.2
python-decouple==3.8
datasets==2.16.1
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1
numpy<2.0
opencv-python<4.10
tensorboard==2.15.1
timm>=0.9.0
scipy>=1.11.0
